weather_bot/requirements.txt
py-clob-client>=0.34.0
httpx>=0.27.0
websockets>=13.0
python-dotenv>=1.0.0
rich>=13.0.0
scipy>=1.12.0
numpy>=1.26.0
python-telegram-bot>=21.0
